requirements.txt
docker==2.0.2
python-dateutil==2.6.0
